backend/requirements.txt
asgiref==3.7.2
Django>=4.2.10
djangorestframework==3.14.0
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2024.1
setuptools==69.0.3
sqlparse==0.4.4
wheel==0.42.0
psycopg2-binary==2.9.9
pillow==10.2.0
django-allauth
dj-rest-auth
python-dotenv
django-cors-headers
drf-yasg
daphne
channels
channels-redis